+++ src/router/router.js
@@ -4,7 +4,7 @@ import TodoList from '../components/TodoList.vue'
 import Notes from '../components/Notes.vue'
 import PetList from '../components/PetList.vue'
 import Width from '../components/Width.vue';
-import Props from '../components/Props.vue';
+import Cities from '../components/Cities.vue';
 
 import VueRouter from 'vue-router'
 
@@ -39,9 +39,9 @@ export default new VueRouter({
             component: Width
         },
         {
-            name: 'props',
-            path: 'props',
-            component: Props
+            name: 'cities',
+            path: 'cities',
+            component: Cities
         }
     ]
 })
